Add files via upload
With the aid of magnetic resonance imaging (MRI), deep learning is utilized to create models for the detection and categorization of brain tumors. This allows for the quick and simple identification of brain tumors.
Machine Learning (ML) algorithms have been successfully applied in the healthcare domain to diagnose diseases. In our work, we show that the use of libraries such as AutoGluon can help to compare the performances of different ML approaches in diagnosing disease for a given dataset with optimal lines of code
This used  Keras, Tensorflow, and OpenCV for Image processing. and Provide Accurate Results